Add GitHub Actions workflow for .NET project (#1)
* Add GitHub Actions workflow for .NET project.

This workflow automates the build and test process for .NET projects on push and pull request events to the main branch. The xUnit tests in the `test/` folder are run after building the dotnet solution for continuous integration testing. Updated suggested GitHub Action to .NET version 9.0 to match the projects in the repository.